protecting api routes

+++ src/components/WorkoutDetails.js
@@ -2,14 +2,22 @@ import React from "react";
 import { formatDistanceToNow } from "date-fns";
 
 import { useWorkoutsContext } from "hooks/useWorkoutsContext";
+import { useAuthContext } from "hooks/useAuthContext";
 import { DELETE_WORKOUT } from "actions";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
 
     const json = await response.json();
